+++ packages/spacecat-shared-gpt-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-gpt-client-v1.5.13](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.5.12...@adobe/spacecat-shared-gpt-client-v1.5.13) (2025-05-26)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#763](https://github.com/adobe/spacecat-shared/issues/763)) ([643b396](https://github.com/adobe/spacecat-shared/commit/643b396130e9144f87099935204927b5522aff68))
+
 # [@adobe/spacecat-shared-gpt-client-v1.5.12](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.5.11...@adobe/spacecat-shared-gpt-client-v1.5.12) (2025-05-19)